Update movie contract guard
